Allow raw pointers in property serialization
