+++ Utils/Utils/Serialization/XMLDeserialize.cpp
@@ -204,6 +204,10 @@ void C_XMLDeserializer::DeserializeProperty(const rttr::property& prop, rttr::va
 			const auto constructPointer = [](const pugi::xml_node& currentNode, const rttr::type& fallbackType) {
 				const pugi::char_t* className	= nullptr;
 				rttr::type			derivedType = fallbackType;
+				if (derivedType.is_pointer())
+				{
+					derivedType = derivedType.get_raw_type();
+				}
 				if (!currentNode.attribute("derivedTypeCast").empty())
 				{
 					className	= currentNode.attribute("derivedTypeCast").as_string();
